fw/apps/system/settings: move Text Size to Display
The text size preference is global and phone-synced, not specific to
notifications, and its placement under Notifications is why users
report the setting as missing. Move the row and its option menu to
Settings -> Display, next to Language. The strings and the
SettingsContentSize mapping are unchanged, so existing translations
keep resolving.
